% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{calculateExpressedRange}
\alias{calculateExpressedRange}
\title{Make a data frame predicting the expression of glycogenes from pseudobulk data}
\usage{
calculateExpressedRange(single_cell.list, reference = available_ranges())
}
\description{
Using the result data from calculatePseudobulks, we can
calculate the set of genes that are predicted to be expressed, and
where their pseudobulk values sit compared to a reference range
}
