% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{calculatePseudobulks}
\alias{calculatePseudobulks}
\title{Trimmed mean pseudobulk and pseudopresence calculation}
\usage{
calculatePseudobulks(count.data, renormalize = FALSE, ...)
}
\arguments{
\item{count.data}{Count data as a matrix, sparse matrix or Seurat object}
}
\value{
Pseudobulks, pseudopresence and cluster sizes
\itemize{
\item pseudobulks.sm - Sparse matrix with pseudobulk values
\item pseudopresence.sm - Sparse matrix with counts of cells that non-zero expression was found in
\item cluster_sizes.vec - Sizes of individual clusters
}
}
\description{
Compute pseudobulks and pseudopresence values for a
count matrix. The pseudobulk is the average expression
count value for a single cluster of cells. The pseudopresence
is the number of cells that a non-zero count was found in.
}
