% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{calculatePseudobulks.Seurat}
\alias{calculatePseudobulks.Seurat}
\title{Trimmed mean pseudobulk and pseudopresence calculation}
\usage{
\method{calculatePseudobulks}{Seurat}(obj.seurat, renormalize = FALSE)
}
\arguments{
\item{obj.seurat}{seurat object with raw counts and cell types as identities}
}
\value{
a list of pseudobulks, pseudopresences and cluster sizes
}
\description{
Compute pseudobulks on seurat object
}
