% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{clr_transform}
\alias{clr_transform}
\title{compute clrs for a vector of values}
\usage{
clr_transform(values)
}
\arguments{
\item{values}{a vector of values}
}
\value{
a vector of clrs
}
\description{
this function determines the clrs of a vector
}
