% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{compute_intersects}
\alias{compute_intersects}
\title{Predicting glycosylation capacity in test data from reference data}
\usage{
compute_intersects(test_data, dynamic_ranges.char)
}
\arguments{
\item{test_data}{an n gene x m sample matrix of values.}

\item{dynamic_ranges.char}{character reference to a stored set of ranges}
}
\value{
the set of prediction statuses computed from applying the expression quantiles from the reference data to each sample in the test data
}
\description{
this function determines whether a gene is expressed in test data by leveraging reference data
}
\examples{

}
