% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{compute_pseudobulk}
\alias{compute_pseudobulk}
\title{trimmed mean pseudobulk and pseudopresence calculation}
\usage{
compute_pseudobulk(obj.seurat)
}
\arguments{
\item{obj.seurat}{seurat object with raw counts and cell types as identities}
}
\value{
a list of pseudobulks, pseudopresences and cluster sizes
}
\description{
computing pseudobulks on seurat object
}
\examples{

}
