% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_glycogenome.R
\name{facet_glycogenome}
\alias{facet_glycogenome}
\title{ggplot2 facet based on glycogenome pathway}
\usage{
facet_glycogenome(
  genes = NULL,
  cols = NULL,
  scales = "free",
  space = "free",
  shrink = TRUE
)
}
\description{
ggplot2 facet based on glycogenome pathway
}
