% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_glycogenes}
\alias{get_glycogenes}
\title{Retrieval of glycogenes from set list.}
\usage{
get_glycogenes(organism = c("human", "mouse"))
}
\arguments{
\item{organism}{a character argument specifying the human or mouse specific gene symbols to return. Default  value is 'human'}
}
\value{
a chqracter vector of glycogenes
}
\description{
this function retrieves the set of glycogenes to subset our reference and input data by
}
\examples{
\dontrun{
genes_human <- get_glycogenes()
genes_mouse <- get_glycogenes('mouse')
}
}
