% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_tissue_results_dpagt1}
\alias{get_tissue_results_dpagt1}
\title{Retrieval of reference gene normalised pseudobulks from PanglaoDB}
\usage{
get_tissue_results_dpagt1(organism = c("human", "mouse"))
}
\arguments{
\item{organism}{a character argument specifying the human or mouse specific reference dataset to return. Default  value is 'human'}
}
\value{
a data.frame of reference gene normalized pseudobulks
}
\description{
this function retrieves the pre-computed reference gene normalised pseudobulks from PanglaoDB
}
\examples{
\dontrun{
reference_data <- get_tissue_results_dpagt1()
reference_data <- get_tissue_results_dpagt1('mouse')
}
}
