% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{getNormalisedRanges}
\alias{getNormalisedRanges}
\title{Predicting dynamic ranges in test data}
\usage{
getNormalisedRanges(input.mat, quantiles = c(0.05, 0.1, 0.25, 0.75, 0.9, 0.95))
}
\arguments{
\item{input.mat}{an n gene x m sample matrix of values}

\item{quantiles}{desired ranges to compute}
}
\value{
the set of ranges computed from applying the specified quantiles to each gene in the test data
}
\description{
this function determines the ranges of a gene in test data
}
