% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{main}
\alias{main}
\title{Predicting glycosylation capacity in bulk data from single cell pseudobulks}
\usage{
main(
  data,
  reference_data = get_tissue_results_dpagt1,
  organism = c("human", "mouse"),
  tissues = NA,
  quantile_cutoff = c(0.25, 0.75),
  reference_gene = c("DPAGT1"),
  genes = get_glycogenes,
  log_transform = c("log", "log1p")
)
}
\arguments{
\item{data}{an n gene x m sample matrix of bulk data values. Expression units can be in either TPM or FPKM}

\item{reference_data}{one of either a) a function returning a data.frame of reference gene normalized single cell pseudobulks, or b) a user supplied data.frame in the same format as that returned by the function in a. Default is (a)}

\item{organism}{an (a) character argument specifying the human or mouse specific datasets and gene symbols to return, or (b) an NA value indicating that user is using own data. Default value is 'human'}

\item{tissues}{one of either a) a character vector of tissues and/or celltypes that the reference data can be filtered for, or b) an NA value indicating that all tissues/celltypes should be used. Default is (b)}

\item{quantile_cutoff}{a numeric vector of values between (0, 1) indicating the quantile to be computed for the normalised pseudobulk distribution of each gene. Default is c(0.25, 0.75)}

\item{reference_gene}{a character vector of reference gene(s) by which to normalise the input data matrix. If more than one reference gene is supplied, the geometric mean of normalized pseudobulks for the genes are computed in each sample. Default is c("DPAGT1")}

\item{genes}{one of either a) a function returning a character vector of glycogenes to be normalised, b) a user supplied character vector of genes, or c) an NA value indicating that all genes in the input data should be normalised. Default is (a)}

\item{log_transform}{a function of a) natural log, b) log1p. Additional log transform functions can be supplied by the user. Default is (a).}
}
\value{
a named list of three data.frames.
normalised_expression are the reference gene normalised values in each bulk input sample.
expression_quantiles are the set of values computed at each specified quantile on the normalised pseudobulk distribution of each gene in the reference data.
predicted_expression are the set of prediction statuses computed from applying the expression quantiles from the reference data to each bulk input sample.
}
\description{
this function determines whether a gene is expressed in bulk data by leveraging single cell reference data from PanglaoDB
}
\examples{
input_data <- as.matrix(read.table(
system.file("extdata", "GT_FPKM.csv", package = "glycoCapacityPredictR", mustWork = TRUE),
header = TRUE, sep = ',', row.names = 1))

head(main(input_data)$normalised_expression)

head(main(input_data,
quantile_cutoff = c(0.95))$normalised_expression)

head(main(input_data,
reference_gene = c("DPAGT1","STT3A"))$normalised_expression)

head(main(input_data,
tissues = 'Peripheral_blood_mononuclear_cells:Dendritic cells')$normalised_expression)

head(main(input_data,
genes = c("XYLT1","XYLT2","STT3B"))$normalised_expression)

head(main(input_data,
log_transform = 'log1p')$normalised_expression)

}
