% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{normalise}
\alias{normalise}
\title{reference gene normalisation}
\usage{
normalise(genes, values, log_transform = T)
}
\arguments{
\item{genes}{a character vector of the sample genes}

\item{values}{a named numeric vector of the counts to be transformed. Names are genes}

\item{log_transform}{boolean indicating if a natural log + 1 transform should be applied}
}
\value{
a data.frame of genes, annotations, and normalised counts
}
\description{
normalising on a reference gene for downstream comparison of glycogene ranges of expression
}
\examples{

}
