% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{normaliseOnSEGs}
\alias{normaliseOnSEGs}
\title{reference gene normalisation wrapper}
\usage{
normaliseOnSEGs(input.mat, genes = row.names(input.mat), log_transform = T)
}
\arguments{
\item{input.mat}{n gene x k sample matrix of suitable counts}

\item{genes}{a named numeric vector of the counts to be transformed. default is rownames of input_mat}

\item{log_transform}{boolean indicating if a natural log + 1 transform should be applied}
}
\value{
a data.frame of genes, annotations, and normalised counts
}
\description{
normalising on a reference gene for downstream comparison of glycogene ranges of expression
}
\examples{

}
