% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{predictMinimalExpression}
\alias{predictMinimalExpression}
\title{single cell gene expression prediction wrapper}
\usage{
predictMinimalExpression(single_cell.list)
}
\arguments{
\item{single_cell.list}{list of pseudobulk sparse matrix and vector of cluster sizes}
}
\value{
an n gene x k cluster matrix of pvalues corresponding to the probability that the gene is expressed in the cluster
}
\description{
predicted expression of a gene in a single cell cluster
}
