% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{predictPseudopresenceStability}
\alias{predictPseudopresenceStability}
\title{single cell pseudopresence prediction wrapper}
\usage{
predictPseudopresenceStability(single_cell.list)
}
\arguments{
\item{single_cell.list}{list of pseudobulk and pseudopresence sparse matrices}
}
\value{
a list of matrices corresponding to  expected mean pseudopresence, expected var pseudopresence and stability pvalue
}
\description{
predicted expression of pseudobulk stability in a single cell cluster
}
