% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{pseudopresence_prediction}
\alias{pseudopresence_prediction}
\title{single cell pseudopresence prediction}
\usage{
pseudopresence_prediction(pseudobulk, pseudopresence = NULL)
}
\arguments{
\item{pseudobulk}{mean expression of the gene in the cluster}

\item{pseudopresence}{number of cells with at least one transcript for the gene in the cluster}
}
\value{
a list of expected mean pseudopresence, expected var pseudopresence and a pvalue corresponding to the probability that the pseudobulk is stable
}
\description{
predicted expression of pseudobulk stability in a single cell cluster
}
