% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{range_generation}
\alias{range_generation}
\title{compute quantiles for a vector of values}
\usage{
range_generation(values, quantiles)
}
\arguments{
\item{values}{a vector of values}

\item{quantiles}{a vector of desired quantiles}
}
\value{
a vector of quantiles
}
\description{
this function determines the quantiles of a vector
}
