% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{transformToCLR}
\alias{transformToCLR}
\title{wrapper for computing clrs on test data}
\usage{
transformToCLR(input.mat)
}
\arguments{
\item{input.mat}{an n gene x k sample matrix of values}
}
\value{
the set of ranges computed from applying the specified quantiles to each gene in the test data
}
\description{
this function determines the clrs of a gene in test data
}
